Create README.md
This library is intended to be the base for a Python based AS4100 Steel Design solver. AS4100 is the current Australian Standard for Structural Steel Design. This library is not intended for people who are not qualified Structural engineers with knowledge of and experience applying this Australian Standard.

It has not been checked / verified except for in-code unit tests. Further testing against other steel design solvers (Excel, software based etc.) will be carried out at some point but has not yet been conducted.

This library is not for commercial use.