Added get_section method to Beam class.
Added a get_section method to the Beam class to support the CodeCheck
objects. Note that this is likely not fully implemented yet and hence
there is 1x test in the test code that is a simple ``assert False`` test
to highlight the fact that there is likely additional work to do here,
even if only for testing.