Refactored get_sections method and added a get_all_sections method.
Refactored get_sections method to allow it to handle the case of
multiple input positions. To support this I needed to refactor the
expected output. This meant that handling the case of an absent position
value became more complex, so added a get_all_sections method to handle
getting the sections without requiring a position.